pyproject.toml
[project]
name = "tmp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy>=2.3.1",
    "pandas>=2.0.0",
    "pyarrow>=10.0.0",
    "huggingface_hub>=0.20.0",
    "python-dotenv>=1.0.0",
    "polars>=1.31.0",
    "requests>=2.31.0",
    "pyyaml>=6.0.0",
    "tqdm>=4.66.0",
    "ruff>=0.12.5",
]
